gate formatting and plurals behind features
