+++ leptos_i18n/src/locale_traits.rs
@@ -1,8 +1,7 @@
+use icu_locid::{LanguageIdentifier, Locale as IcuLocale};
+use leptos_router::ChooseView;
 use std::hash::Hash;
 use std::str::FromStr;
-
-use icu::locid;
-use leptos_router::ChooseView;
 
 use crate::langid::{convert_vec_str_to_langids_lossy, filter_matches, find_match};
 
@@ -15,8 +14,8 @@ pub trait Locale<L: Locale = Self>:
     + Clone
     + Copy
     + FromStr
-    + AsRef<locid::LanguageIdentifier>
-    + AsRef<locid::Locale>
+    + AsRef<LanguageIdentifier>
+    + AsRef<IcuLocale>
     + AsRef<str>
     + AsRef<L>
     + std::fmt::Display
@@ -39,10 +38,10 @@ pub trait Locale<L: Locale = Self>:
     fn as_str(self) -> &'static str;
 
     /// Return a static reference to a icu `Locale`
-    fn as_icu_locale(self) -> &'static locid::Locale;
+    fn as_icu_locale(self) -> &'static IcuLocale;
 
     /// Return a static reference to a `LanguageIdentifier`
-    fn as_langid(self) -> &'static locid::LanguageIdentifier {
+    fn as_langid(self) -> &'static LanguageIdentifier {
         Locale::as_icu_locale(self).as_ref()
     }
 
@@ -59,7 +58,7 @@ pub trait Locale<L: Locale = Self>:
     /// Given a langid, return a Vec of suitables `Locale` sorted in compatibility (first one being the best match).
     ///
     /// This function does not fallback to default if no match is found.
-    fn find_matchs<T: AsRef<locid::LanguageIdentifier>>(langid: T) -> Vec<Self> {
+    fn find_matchs<T: AsRef<LanguageIdentifier>>(langid: T) -> Vec<Self> {
         let matches: Vec<L> =
             filter_matches(std::slice::from_ref(langid.as_ref()), Self::get_all());
         matches.into_iter().map(Self::from_base_locale).collect()
